独立出Design控件，同时增加预览功能
