+++ package.json
@@ -54,7 +54,10 @@
       ]
     },
     "parserOptions": {
-      "parser": "@typescript-eslint/parser"
+      "parser": "@typescript-eslint/parser",
+      "ecmaFeatures": {
+        "jsx": false
+      }
     },
     "overrides": [
       {
